Moving API access function to main.ts.
