+++ src/main.ts
@@ -677,3 +677,13 @@ class LocalRestApiSettingTab extends PluginSettingTab {
     }
   }
 }
+
+export const getAPI = (
+  app: App,
+  manifest: PluginManifest
+): LocalRestApiPublicApi | undefined => {
+  const plugin = app.plugins.plugins["obsidian-local-rest-api"];
+  if (plugin) {
+    return (plugin as unknown as LocalRestApi).getPublicApi(manifest);
+  }
+};
